% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genePlot.R
\name{genePlot}
\alias{genePlot}
\title{Plot splicing ratios for each exon in a given gene}
\usage{
genePlot(
  data,
  gene,
  condition = NULL,
  ylab = "exon relative abundance",
  xlab = "Condition"
)
}
\arguments{
\item{data}{expression dataset. Must be of class `RangedSummarizedExperiment`.}

\item{condition}{a `factor` representing the condition of each individual.}

\item{ylab}{y-axis label. Optional.}

\item{xlab}{x-axix label. Optional.}
}
\description{
Plot splicing ratios for each exon in a given gene
}
\examples{
data(adipose)
genePlot(adipose.chr16, gene="ENSG00000197165.10", condition="inv16p11.2")

}
