% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotAllIso.R
\name{plotAllIso}
\alias{plotAllIso}
\title{Plot proportional exon expression}
\usage{
plotAllIso(gene, data, group, jitterWidth = 0, inds, isos)
}
\arguments{
\item{gene}{name of the gene to plot.}

\item{data}{expression dataset. Can be of class `DEXSeqDataSet`, `SummarizedExperiment` or
`RangedSummarizedExperiment`. A `matrix` or `data.frame` is also accepted.}

\item{group}{a `factor` representing the condition of each individual.
Required if `data` is a `matrix` or `data.fame`, ignored otherwise.}

\item{jitterWidth}{amount of graphical jitter to use in the plot.}

\item{inds}{`character` or `numeric` vector. Plot only the specified individuals (samples).}

\item{isos}{`character` or `numeric` vector. Plot only the specified exons.
#' @examples
data(adipose)
plotAllIso("ENSG00000003249.13", adipose.chr16, "inv16p11.2")

data(YRI)
plotAllIso("ENSG00000215915", YRI, attr(YRI, "gender"))}
}
\description{
Plot proportional exon expression
}
