% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotAllIsoRSE.R
\name{plotAllIsoRSE}
\alias{plotAllIsoRSE}
\title{Plot proportional exon expression stored in a RangeSummarizedExperiment object}
\usage{
plotAllIsoRSE(gene, data, group, jitterWidth = 0, inds, isos)
}
\arguments{
\item{gene}{name of the gene to plot.}

\item{data}{expression dataset. Must be of class `RangedSummarizedExperiment`.}

\item{group}{a `factor` representing the condition of each individual.}

\item{jitterWidth}{amount of graphical jitter to use in the plot.}

\item{inds}{`character` or `numeric` vector. Plot only the specified individuals (samples).}

\item{isos}{`character` or `numeric` vector. Plot only the specified exons.}
}
\description{
Plot proportional exon expression stored in a RangeSummarizedExperiment object
}
\examples{
data(adipose)
plotAllIsoRSE("ENSG00000003249.13", adipose.chr16, "inv16p11.2")

}
