% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotSplicingRatios.R
\name{plotSplicingRatios}
\alias{plotSplicingRatios}
\title{Plot splicing ratios of exon expression stored in a RangeSummarizedExperiment object}
\usage{
plotSplicingRatios(data, gene, condition, jitterWidth = 0, inds, exons)
}
\arguments{
\item{data}{expression dataset. Must be of class `RangedSummarizedExperiment`.}

\item{gene}{name of the gene to plot.}

\item{condition}{a `factor` representing the condition of each individual.}

\item{jitterWidth}{amount of graphical jitter to use in the plot.}

\item{inds}{`character` or `numeric` vector. Plot only the specified individuals (samples).}

\item{exons}{`character` or `numeric` vector. Plot only the specified exons.}
}
\description{
Plot splicing ratios of exon expression stored in a RangeSummarizedExperiment object
}
\examples{
data(adipose)
plotSplicingRatios("ENSG00000003249.13", adipose.chr16, "inv16p11.2")

}
