% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotTernary.R
\name{plotTernary}
\alias{plotTernary}
\title{Plot proportional exon expression}
\usage{
plotTernary(x, gene, group, geneCol = 1, transCol = 2, transcripts, ...)
}
\arguments{
\item{x}{expression dataset. Can be of class `DEXSeqDataSet`, `SummarizedExperiment` or
`RangedSummarizedExperiment`. A `matrix` or `data.frame` is also accepted.}

\item{gene}{name of the gene to plot.}

\item{group}{a `factor` representing the condition of each individual.
Required if `data` is a `matrix` or `data.fame`, ignored otherwise.}

\item{geneCol}{`character` or `numeric` vector specifying the columns of `x` with that contain the gene names.
Required if `x` is a `matrix` or `data.fame`, ignored otherwise.}

\item{transCol}{`character` or `numeric` vector specifying the columns of `x` with that contain the transcript names.
Required if `x` is a `matrix` or `data.fame`, ignored otherwise.}

\item{transcripts}{`numeric` vector of length 3 with the indices of the transcripts to ploth.}
}
\description{
Plot proportional exon expression
}
\examples{
data(adipose)
plotTernary(adipose.chr16, "ENSG00000003249.13", transcripts = c(2, 4, 7))

data(YRI)
plotTernary(YRI, "ENSG00000160741")
}
