% --- rasp-internal.Rd ---

\name{rasp-internal}
\alias{rasp-internal}
\alias{hellingerDist}


\title{Internal rasp functions}
\description{Internal rasp functions}
\usage{
hellingerDist(x1, x2)
testRaspmc(x, group, test, type="median", Nperms, ...)
}

\details{These are not to be called by the user}
\keyword{internal}
