% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topGenes.R
\name{topGenes}
\alias{topGenes}
\title{Get the top ranking genes.}
\usage{
topGenes(
  x,
  n = 6L,
  sort.by = "pvalue",
  pval.adjust.cutoff = 1,
  print = TRUE,
  ...
)
}
\arguments{
\item{x}{object returned by `rasp`.}

\item{n}{number of top genes to return.}

\item{sort.by}{column to be used as a sorting key.}

\item{pval.adjust.cutoff}{return only genes with a `padjust` value lower than this.}

\item{print}{whether to print the output, in addition to only returning.}

\item{...}{additional arguments passed to `print.data.frame`.}
}
\value{
a `data.frame`
}
\description{
By default, `topGenes` will return *and* print the 6 genes with best p-value.
}
