% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rasp.R
\name{rasp}
\alias{rasp}
\title{rasp}
\usage{
rasp(
  formula,
  x,
  group,
  expressionCols,
  geneidCol,
  filterInd = 0.1,
  filterExon = 0.05,
  transform = "none",
  cores = 1,
  ...
)
}
\arguments{
\item{formula}{a `formula` with a symbolic description of the model to be fitted.}

\item{x}{expression dataset. Can be of class `DEXSeqDataSet`, `SummarizedExperiment` or
`RangedSummarizedExperiment`. A `matrix` or `data.frame` is also accepted.}

\item{group}{a `factor` representing the condition of each individual.
Required if `x` is a `matrix` or `data.fame`, ignored otherwise.}

\item{expressionCols}{`character` or `numeric` vector specifying the columns of `x` with that contain
the expression data (i.e. the different samples).
Required if `x` is a `matrix` or `data.fame`, ignored otherwise.}

\item{geneidCol}{name or index of the column of `x` encoding the gene IDs.
Required if `x` is a `matrix` or `data.fame`, ignored otherwise.}

\item{filterInd}{filter individuals (samples) with mean relative frequency < `filterInd`.}

\item{filterExon}{filter exons with mean relative frequency < `filterFreq`.}

\item{transform}{argument for `mlm`. Can be any of `none`, `sqrt` or `log`.}

\item{cores}{`integer`. Number of parallel processes to use.}

\item{...}{additional arguments passed to `mlm`.}
}
\value{
A `rasp` class `matrix` containing the p-values (also adjusted for multiple testing).
  `NA` will be returned for single exon genes and genes not passing filtering criteria.
}
\description{
Differential alternative splicing test.
}
\details{
Test for differential alternative splicing across two or more conditions.
}
\examples{
data(adipose)
rasp(~ inv16p11.2, adipose.chr16)

data("YRI")
rasp(x = YRI, group = attr(YRI, "gender"), expressionCols = 3:71, geneidCol = "gene_id")
}
